requirements.txt
torch
torchvision
timm
scikit-learn
matplotlib
tqdm
pillow
numpy
pandas
